+++ lib/appium_lib/core/android/uiautomator2/bridge.rb
@@ -8,6 +8,9 @@ module Appium
           def self.for(target)
             target.extend Appium::Android::Device
             ::Appium::Core::Android::SearchContext.extend
+
+            ::Appium::Core::Android::Touch.extend_touch_actions
+            ::Appium::Core::Android::Touch.extend_multi_touch
           end
         end
       end
