+++ lib/appium_lib/core/android_uiautomator2.rb
@@ -1,4 +1,5 @@
 # loaded in common/driver.rb
 require_relative 'android/search_context'
 require_relative 'android/device'
+require_relative 'android/touch'
 require_relative 'android/uiautomator2/bridge'
