+++ lib/appium_lib/core/ios.rb
@@ -4,3 +4,4 @@ require_relative 'ios/device'
 
 require_relative 'ios/uiautomation/patch'
 require_relative 'ios/uiautomation/bridge'
+require_relative 'ios/uiautomation/touch'
