+++ lib/appium_lib/core/ios/xcuitest/bridge.rb
@@ -10,6 +10,9 @@ module Appium
             Core::Ios::Xcuitest::SearchContext.extend
             target.extend Appium::Ios::Device
             target.extend Appium::Ios::Xcuitest::Device
+
+            ::Appium::Core::Ios::Touch.extend_touch_actions
+            ::Appium::Core::Ios::Touch.extend_multi_touch
           end
         end
       end
