+++ lib/appium_lib/core/ios_xcuitest.rb
@@ -5,3 +5,4 @@ require_relative 'ios/device'
 require_relative 'ios/xcuitest/search_context'
 require_relative 'ios/xcuitest/device'
 require_relative 'ios/xcuitest/bridge'
+require_relative 'ios/xcuitest/touch'
